Create Leap_Years.c
In the Gregorian calendar, a leap year is created by extending February to 29 days in order to keep the calendar year synchronized with the astronomical year. These longer years occur in years which are multiples of 4, with the exception of centennial years that aren’t multiples of 400.

Write a program to print all the leap years from the year 1800 up to and including 2400.